refactor: abstract delete category mutation
